Renaming templateMap to templates
