+++ packages/create-cloudflare/src/templates.ts
