+++ packages/create-cloudflare/templates/astro/c3.ts
@@ -4,7 +4,7 @@ import { runCommand, runFrameworkGenerator } from "helpers/command";
 import { compatDateFlag } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
 import type { C3Context } from "types";
-import { TemplateConfig } from "../../src/templateMap";
+import { TemplateConfig } from "../../src/templates";
 
 const { npx } = detectPackageManager();
