+++ packages/create-cloudflare/templates/solid/c3.ts
@@ -4,7 +4,7 @@ import { installPackages, runFrameworkGenerator } from "helpers/command";
 import { compatDateFlag } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
 import type { C3Context } from "types";
-import { TemplateConfig } from "../../src/templateMap";
+import { TemplateConfig } from "../../src/templates";
 
 const { npm } = detectPackageManager();
