+++ packages/create-cloudflare/templates/svelte/c3.ts
@@ -7,7 +7,7 @@ import { detectPackageManager } from "helpers/packages";
 import { platformInterface } from "./templates";
 import type * as recast from "recast";
 import type { C3Context } from "types";
-import { TemplateConfig } from "../../src/templateMap";
+import { TemplateConfig } from "../../src/templates";
 
 const { npm } = detectPackageManager();
